# docs/swagger.yaml
basePath: /api/v1
definitions:
  domain.CreateTeamRequest:
    properties:
      teamName:
        type: string
    type: object
  domain.Folder:
    properties:
      createdAt:
        type: string
      folderId:
        type: integer
      name:
        type: string
      notes:
        items:
          $ref: '#/definitions/domain.Note'
        type: array
      ownerId:
        type: integer
      updatedAt:
        type: string
    type: object
  domain.ImportResult:
    properties:
      errors:
        items:
          type: string
        type: array
      failed:
        type: integer
      succeeded:
        type: integer
    type: object
  domain.LoginRequest:
    properties:
      email:
        type: string
      password:
        type: string
    type: object
  domain.Note:
    properties:
      content:
        type: string
      createdAt:
        type: string
      folderId:
        type: integer
      noteId:
        type: integer
      title:
        type: string
      updatedAt:
        type: string
    type: object
  domain.RegisterRequest:
    properties:
      email:
        type: string
      password:
        type: string
      role:
        $ref: '#/definitions/domain.Role'
      username:
        type: string
    type: object
  domain.Role:
    enum:
    - manager
    - member
    type: string
    x-enum-varnames:
    - RoleManager
    - RoleMember
  domain.Team:
    properties:
      createdAt:
        type: string
      managers:
        items:
          $ref: '#/definitions/domain.User'
        type: array
      members:
        items:
          $ref: '#/definitions/domain.User'
        type: array
      teamId:
        type: integer
      teamName:
        type: string
      updatedAt:
        type: string
    type: object
  domain.User:
    properties:
      createdAt:
        type: string
      email:
        type: string
      role:
        $ref: '#/definitions/domain.Role'
      updatedAt:
        type: string
      userId:
        type: integer
      username:
        type: string
    type: object
host: localhost:3000
info:
  contact:
    email: support@swagger.io
    name: API Support
    url: http://www.swagger.io/support
  description: This is a sample server for a microservices challenge.
  license:
    name: Apache 2.0
    url: http://www.apache.org/licenses/LICENSE-2.0.html
  termsOfService: http://swagger.io/terms/
  title: Seta Training API
  version: "1.0"
paths:
  /assets/{type}/{id}/share:
    post:
      consumes:
      - application/json
      description: Share a folder or note with another user
      parameters:
      - description: Asset Type (folder/note)
        in: path
        name: type
        required: true
        type: string
      - description: Asset ID
        in: path
        name: id
        required: true
        type: integer
      - description: Sharing details
        in: body
        name: request
        required: true
        schema:
          additionalProperties: true
          type: object
      produces:
      - application/json
      responses:
        "200":
          description: OK
        "400":
          description: Bad Request
          schema:
            additionalProperties:
              type: string
            type: object
        "403":
          description: Forbidden
          schema:
            additionalProperties:
              type: string
            type: object
      security:
      - Bearer: []
      summary: Share an asset
      tags:
      - assets
  /assets/{type}/{id}/share/{targetUserId}:
    delete:
      description: Revoke a user's access to a folder or note
      parameters:
      - description: Asset Type (folder/note)
        in: path
        name: type
        required: true
        type: string
      - description: Asset ID
        in: path
        name: id
        required: true
        type: integer
      - description: User ID to revoke
        in: path
        name: targetUserId
        required: true
        type: integer
      responses:
        "200":
          description: OK
        "400":
          description: Bad Request
          schema:
            additionalProperties:
              type: string
            type: object
        "403":
          description: Forbidden
          schema:
            additionalProperties:
              type: string
            type: object
      security:
      - Bearer: []
      summary: Revoke access to an asset
      tags:
      - assets
  /assets/folders:
    post:
      consumes:
      - application/json
      description: Create a new digital folder for assets
      parameters:
      - description: Folder details
        in: body
        name: request
        required: true
        schema:
          additionalProperties:
            type: string
          type: object
      produces:
      - application/json
      responses:
        "201":
          description: Created
          schema:
            $ref: '#/definitions/domain.Folder'
        "400":
          description: Bad Request
          schema:
            additionalProperties:
              type: string
            type: object
      security:
      - Bearer: []
      summary: Create a new folder
      tags:
      - assets
  /assets/folders/{id}:
    get:
      description: Retrieve a folder and its notes, checking access permissions
      parameters:
      - description: Folder ID
        in: path
        name: id
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/domain.Folder'
        "400":
          description: Bad Request
          schema:
            additionalProperties:
              type: string
            type: object
        "403":
          description: Forbidden
          schema:
            additionalProperties:
              type: string
            type: object
      security:
      - Bearer: []
      summary: Get folder details
      tags:
      - assets
  /assets/folders/{id}/notes:
    post:
      consumes:
      - application/json
      description: Create a new note within a specified folder
      parameters:
      - description: Folder ID
        in: path
        name: id
        required: true
        type: integer
      - description: Note details
        in: body
        name: request
        required: true
        schema:
          additionalProperties:
            type: string
          type: object
      produces:
      - application/json
      responses:
        "201":
          description: Created
          schema:
            $ref: '#/definitions/domain.Note'
        "400":
          description: Bad Request
          schema:
            additionalProperties:
              type: string
            type: object
        "403":
          description: Forbidden
          schema:
            additionalProperties:
              type: string
            type: object
      security:
      - Bearer: []
      summary: Create a note in a folder
      tags:
      - assets
  /assets/notes/{id}:
    get:
      description: Retrieve a specific note, checking access permissions
      parameters:
      - description: Note ID
        in: path
        name: id
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/domain.Note'
        "400":
          description: Bad Request
          schema:
            additionalProperties:
              type: string
            type: object
        "403":
          description: Forbidden
          schema:
            additionalProperties:
              type: string
            type: object
      security:
      - Bearer: []
      summary: Get note details
      tags:
      - assets
    put:
      consumes:
      - application/json
      description: Update the title or content of an existing note
      parameters:
      - description: Note ID
        in: path
        name: id
        required: true
        type: integer
      - description: Note updates
        in: body
        name: request
        required: true
        schema:
          additionalProperties:
            type: string
          type: object
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/domain.Note'
        "400":
          description: Bad Request
          schema:
            additionalProperties:
              type: string
            type: object
        "403":
          description: Forbidden
          schema:
            additionalProperties:
              type: string
            type: object
      security:
      - Bearer: []
      summary: Update a note
      tags:
      - assets
  /auth/login:
    post:
      consumes:
      - application/json
      description: Authenticate a user and return a JWT token
      parameters:
      - description: Login Credentials
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/domain.LoginRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            additionalProperties:
              type: string
            type: object
        "401":
          description: Unauthorized
          schema:
            additionalProperties:
              type: string
            type: object
      summary: Login a user
      tags:
      - auth
  /auth/register:
    post:
      consumes:
      - application/json
      description: Create a new user with a specific role
      parameters:
      - description: Registration Info
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/domain.RegisterRequest'
      produces:
      - application/json
      responses:
        "201":
          description: Created
          schema:
            $ref: '#/definitions/domain.User'
        "400":
          description: Bad Request
          schema:
            additionalProperties:
              type: string
            type: object
      summary: Register a new user
      tags:
      - auth
  /teams:
    post:
      consumes:
      - application/json
      description: Allows a manager to create a new team
      parameters:
      - description: Team details
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/domain.CreateTeamRequest'
      produces:
      - application/json
      responses:
        "201":
          description: Created
          schema:
            $ref: '#/definitions/domain.Team'
        "400":
          description: Bad Request
          schema:
            additionalProperties:
              type: string
            type: object
      security:
      - Bearer: []
      summary: Create a new team
      tags:
      - teams
  /teams/{id}/managers:
    post:
      consumes:
      - application/json
      description: Allows the main manager to add another manager to the team
      parameters:
      - description: Team ID
        in: path
        name: id
        required: true
        type: integer
      - description: User ID
        in: body
        name: request
        required: true
        schema:
          additionalProperties:
            type: integer
          type: object
      produces:
      - application/json
      responses:
        "200":
          description: OK
        "400":
          description: Bad Request
          schema:
            additionalProperties:
              type: string
            type: object
      security:
      - Bearer: []
      summary: Add a manager to a team
      tags:
      - teams
  /teams/{id}/managers/{userId}:
    delete:
      description: Allows the main manager to remove another manager from the team
      parameters:
      - description: Team ID
        in: path
        name: id
        required: true
        type: integer
      - description: User ID to remove
        in: path
        name: userId
        required: true
        type: integer
      responses:
        "200":
          description: OK
        "400":
          description: Bad Request
          schema:
            additionalProperties:
              type: string
            type: object
      security:
      - Bearer: []
      summary: Remove a manager from a team
      tags:
      - teams
  /teams/{id}/members:
    post:
      consumes:
      - application/json
      description: Allows a team manager to add a member to their team
      parameters:
      - description: Team ID
        in: path
        name: id
        required: true
        type: integer
      - description: User ID
        in: body
        name: request
        required: true
        schema:
          additionalProperties:
            type: integer
          type: object
      produces:
      - application/json
      responses:
        "200":
          description: OK
        "400":
          description: Bad Request
          schema:
            additionalProperties:
              type: string
            type: object
      security:
      - Bearer: []
      summary: Add a member to a team
      tags:
      - teams
  /teams/{id}/members/{userId}:
    delete:
      description: Allows a team manager to remove a member from their team
      parameters:
      - description: Team ID
        in: path
        name: id
        required: true
        type: integer
      - description: User ID to remove
        in: path
        name: userId
        required: true
        type: integer
      responses:
        "200":
          description: OK
        "400":
          description: Bad Request
          schema:
            additionalProperties:
              type: string
            type: object
      security:
      - Bearer: []
      summary: Remove a member from a team
      tags:
      - teams
  /users:
    get:
      description: Retrieve a list of all registered users
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            items:
              $ref: '#/definitions/domain.User'
            type: array
        "500":
          description: Internal Server Error
          schema:
            additionalProperties:
              type: string
            type: object
      security:
      - Bearer: []
      summary: Get all users
      tags:
      - users
  /users/import:
    post:
      consumes:
      - multipart/form-data
      description: Concurrently create users from an uploaded CSV file
      parameters:
      - description: 'CSV file with columns: username, email, password, role'
        in: formData
        name: file
        required: true
        type: file
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/domain.ImportResult'
        "400":
          description: Bad Request
          schema:
            additionalProperties:
              type: string
            type: object
      security:
      - Bearer: []
      summary: Bulk import users from CSV
      tags:
      - users
securityDefinitions:
  Bearer:
    in: header
    name: Authorization
    type: apiKey
swagger: "2.0"